correção de tipo

+++ src/Solucao/solucao.hpp
@@ -12,7 +12,7 @@ struct Solucao {
     void exibe_solucao();
     void add_no(double x);
     void calcula_valor_obj();
-    int dist(int i,int j);
+    double dist(int i,int j);
 };
